[AVR] Implemented AVRRegisterInfo::getFrameRegister()
This function gets the register used to store a pointer to the stack
frame.

I believe that if a function requires a dedicated frame pointer, that
pointer is Y (to conform to the AVR-GCC ABI). Otherwise, the stack
pointer register is used.

Closes #22

See https://gcc.gnu.org/wiki/avr-gcc for more details.