[AVR] Removed MUL16RdRr instruction
It had the exact same signature as `MUL Rd, Rr`, and the same machine
encoding, although implemented incorrectly. When the asm parser
encountered `MUL Rd, Rr`, it would interpret it as MUL16RdRr, causing
encoding errors.

The instruction is redundant, and all references to it are now replaced
with MULRdRr.