[AVR] Fixed PCRel fixup code
We would be checking the rightshifted value if it fit in an integer of
`n` bits.

This meant that it would incorrectly say that a 7 bit value
could not fit into the 7 bit field. It also would always cause an error
for negative labels (that occur previously in the program with respect
to the location of the fixup in memory).

This is now fixed.