[AVR] Refactor multiplication instructions
The mul instructions are closer to the UMUL_LOHI class of operands. Guide the
initial lowering to prefer those operands in AVRISelLowering. Use custom
selection code in AVRISelDAGToDAG to move the result into the target registers.
Finally, clear the zero_reg R1, using a custom inserter to inject a clear
instruction right after the MOVs.