[AVR] Support PC-relative targets with negative values
Beforehand, the value was not correctly adjusted, leading to
negative values being used as seeminly random positive values.

The main culprit of this was the fact that LLVM called
`adjustFixupValue()` on the fixup value before passing it to
`applyFixup()`, but we were adjusting it a second time, giving it a bad
value.