[AVR] Fixed fixup bug
When a fixup had an offset, the mask that gets OR'd with the target
data needed to be bigger.

The problem was that we were assuming the size of the fixup was the
number of bytes that need to be written, but this is wrong - the offset
needs to be considered.