[AVR] avrlit: Add on-target testing framework
This adds two things:

libavrlit: A small testing framework for the AVR, built using
avr-gcc. It reports results to a TTY using a format similar to llvm-lit.

llvm-avrlit: A test driver script. It builds an AVR executable from
code lowered by our backend and libavrlit. It uploads and runs the
test executable on a test board (Arduino Leonardo et al.) and collects
the results.

The on-target tests are only enabled if the AVRLIT_PORT environment
variable is set. This is checked in test/CodeGen/AVR/lit.local.cfg.