[AVR] Fixed 'return.ll' codegen test
Also added exception so that retvals are only reversed when the calling
conv is not a runtime library function.

The update to LLVM master broke it.

The test is incorrect - we should only be spilling registers if they
are clobbered. `MOVW` does not modify the source registers, and so they
should not be spilled.