[AVR] Changed syntax of negated lo8,hi8 modifiers
Previously they were lo8(-value). This is equivalent to taking the low
8 bits of the value and negating it. The old syntax does not read this
way. It looks as if it is negating the value and taking the low 8-bits
of that.

This also falls through when adding a constant offset - how to we
display the low 8-bits of a symbol plus a constant - `lo8(-mysym +
100)`, `lo8(-(mysym + 100))`?

It is now changed to:
`-lo8(val)` and
`-lo8(val + c)`

This commit also removed a garbage file, and implements the 'ms8' fixup.