[AVR][AsmParser] Fixed parameters to peekTokens()
I reverted my original peekTokens implementation, as I had modified it
and upstreamed it to LLVM. This fixes the function call. The reason
this still compiled is that the `2` was being interpreted as the
boolean `ShouldSkipSpace` parameter.

Also updated function call in AVRInstrInfo.cpp to updated signature.
Also fixed a bunch of switch statement warnings.